transferi as funcoes para cadastrar usuario para o SignUpForm
